DEVELOPMENT - converted all js files to typescript
