+++ gatsby-node.js
@@ -1,87 +1,87 @@
-const _ = require('lodash')
-const path = require('path')
-const { createFilePath } = require('gatsby-source-filesystem')
-const { fmImagesToRelative } = require('gatsby-remark-relative-images')
+const _ = require("lodash");
+const path = require("path");
+const { createFilePath } = require("gatsby-source-filesystem");
+const { fmImagesToRelative } = require("gatsby-remark-relative-images");
 
 exports.createPages = ({ actions, graphql }) => {
-  const { createPage } = actions
+    const { createPage } = actions;
 
-  return graphql(`
-    {
-      allMarkdownRemark(limit: 1000) {
-        edges {
-          node {
-            id
-            fields {
-              slug
+    return graphql(`
+        {
+            allMarkdownRemark(limit: 1000) {
+                edges {
+                    node {
+                        id
+                        fields {
+                            slug
+                        }
+                        frontmatter {
+                            tags
+                            templateKey
+                        }
+                    }
+                }
             }
-            frontmatter {
-              tags
-              templateKey
-            }
-          }
         }
-      }
-    }
-  `).then((result) => {
-    if (result.errors) {
-      result.errors.forEach((e) => console.error(e.toString()))
-      return Promise.reject(result.errors)
-    }
+    `).then((result) => {
+        if (result.errors) {
+            result.errors.forEach((e) => console.error(e.toString()));
+            return Promise.reject(result.errors);
+        }
 
-    const posts = result.data.allMarkdownRemark.edges
+        const posts = result.data.allMarkdownRemark.edges;
 
-    posts.forEach((edge) => {
-      const id = edge.node.id
-      createPage({
-        path: edge.node.fields.slug,
-        tags: edge.node.frontmatter.tags,
-        component: path.resolve(
-          `src/templates/${String(edge.node.frontmatter.templateKey)}.js`
-        ),
-        // additional data can be passed via context
-        context: {
-          id,
-        },
-      })
-    })
+        posts.forEach((edge) => {
+            const id = edge.node.id;
+            createPage({
+                path: edge.node.fields.slug,
+                tags: edge.node.frontmatter.tags,
+                component: path.resolve(
+                    `src/templates/${String(edge.node.frontmatter.templateKey)}.tsx`
+                ),
+                // additional data can be passed via context
+                context: {
+                    id,
+                },
+            });
+        });
 
-    // Tag pages:
-    let tags = []
-    // Iterate through each post, putting all found tags into `tags`
-    posts.forEach((edge) => {
-      if (_.get(edge, `node.frontmatter.tags`)) {
-        tags = tags.concat(edge.node.frontmatter.tags)
-      }
-    })
-    // Eliminate duplicate tags
-    tags = _.uniq(tags)
+        // Tag pages:
+        let tags = [];
+        // Iterate through each post, putting all found tags into `tags`
+        posts.forEach((edge) => {
+            if (_.get(edge, `node.frontmatter.tags`)) {
+                tags = tags.concat(edge.node.frontmatter.tags);
+            }
+        });
+        // Eliminate duplicate tags
+        tags = _.uniq(tags);
 
-    // Make tag pages
-    tags.forEach((tag) => {
-      const tagPath = `/tags/${_.kebabCase(tag)}/`
+        // Make tag pages
+        tags.forEach((tag) => {
+            const tagPath = `/tags/${_.kebabCase(tag)}/`;
 
-      createPage({
-        path: tagPath,
-        component: path.resolve(`src/templates/tags.js`),
-        context: {
-          tag,
-        },
-      })
-    })
-  })
-}
+            createPage({
+                path: tagPath,
+                component: path.resolve(`src/templates/tags.tsx`),
+                context: {
+                    tag,
+                },
+            });
+        });
+    });
+};
 
 exports.onCreateNode = ({ node, actions, getNode }) => {
-  const { createNodeField } = actions
-  fmImagesToRelative(node) // convert image paths for gatsby images
+    const { createNodeField } = actions;
+    fmImagesToRelative(node); // convert image paths for gatsby images
 
-  if (node.internal.type === `MarkdownRemark`) {
-    const value = createFilePath({ node, getNode })
-    createNodeField({
-      name: `slug`,
-      node,
-      value,
-    })
-  }
-}
+    if (node.internal.type === `MarkdownRemark`) {
+        const value = createFilePath({ node, getNode });
+        createNodeField({
+            name: `slug`,
+            node,
+            value,
+        });
+    }
+};
