+++ src/components/Content.tsx
@@ -2,15 +2,15 @@ import React from "react";
 
 interface ContentProps {
     content: any;
-    className: string;
+    cssClassName?: string;
 }
 
-export const HTMLContent: React.FC<ContentProps> = ({ className, content }) => (
-    <div className={className} dangerouslySetInnerHTML={{ __html: content }} />
+export const HTMLContent: React.FC<ContentProps> = ({ cssClassName, content }) => (
+    <div className={cssClassName} dangerouslySetInnerHTML={{ __html: content }} />
 );
 
-const Content: React.FC<ContentProps> = ({ className, content }) => (
-    <div className={className}>{content}</div>
+const Content: React.FC<ContentProps> = ({ cssClassName, content }) => (
+    <div className={cssClassName}>{content}</div>
 );
 
 export default Content;
