+++ src/components/pricing.tsx
@@ -1,0 +1,35 @@
+import React from "react";
+
+interface PricingProps {
+    data: Array<{
+        plan: string;
+        price: string | number;
+        description: string;
+        items: Array<any>;
+    }>;
+}
+
+const Pricing: React.FC<PricingProps> = ({ data }) => (
+    <div className="columns">
+        {data.map((price) => (
+            <div key={price.plan} className="column">
+                <section className="section">
+                    <h4 className="has-text-centered has-text-weight-semibold">{price.plan}</h4>
+                    <h2 className="is-size-1 has-text-weight-bold has-text-primary has-text-centered">
+                        ${price.price}
+                    </h2>
+                    <p className="has-text-weight-semibold">{price.description}</p>
+                    <ul>
+                        {price.items.map((item) => (
+                            <li key={item} className="is-size-5">
+                                {item}
+                            </li>
+                        ))}
+                    </ul>
+                </section>
+            </div>
+        ))}
+    </div>
+);
+
+export default Pricing;
