+++ src/interfaces/image-info.ts
@@ -1,0 +1,8 @@
+interface ImageInfo {
+    alt: string;
+    childImageSharp?: any;
+    image: any;
+    style?: any;
+}
+
+export default ImageInfo;
