+++ src/interfaces/testimonial.ts
@@ -1,0 +1,6 @@
+interface Testimonial {
+    author: string;
+    quote: string;
+}
+
+export default Testimonial;
