+++ src/pages/contact/examples.js
@@ -1,6 +1,6 @@
 import React from "react";
 import Link from "gatsby-link";
-import Layout from "../../components/Layout.tsx";
+import Layout from "../../components/layout.tsx";
 
 export default class Index extends React.Component {
     render() {
