+++ src/pages/contact/file-upload.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { navigate } from "gatsby-link";
-import Layout from "../../components/Layout.tsx";
+import Layout from "../../components/layout.tsx";
 
 function encode(data) {
     const formData = new FormData();
