+++ src/pages/contact/thanks.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Layout from "../../components/Layout.tsx";
+import Layout from "../../components/layout.tsx";
 
 export default () => (
     <Layout>
