+++ src/templates/blog-post.tsx
@@ -1,0 +1,98 @@
+import React from "react";
+import { kebabCase } from "lodash";
+import { Helmet } from "react-helmet";
+import { graphql, Link } from "gatsby";
+import Layout from "../components/layout";
+import { HTMLContent } from "../components/content";
+
+interface BlogPostTemplateProps {
+    content: React.FC;
+    contentComponent?: typeof HTMLContent;
+    description: string;
+    tags: Array<string>;
+    title: string;
+    helmet?: any;
+}
+
+export const BlogPostTemplate: React.FC<BlogPostTemplateProps> = ({
+    content,
+    contentComponent,
+    description,
+    tags,
+    title,
+    helmet,
+}) => {
+    const PostContent = contentComponent || HTMLContent;
+
+    return (
+        <section className="section">
+            {helmet || ""}
+            <div className="container content">
+                <div className="columns">
+                    <div className="column is-10 is-offset-1">
+                        <h1 className="title is-size-2 has-text-weight-bold is-bold-light">
+                            {title}
+                        </h1>
+                        <p>{description}</p>
+                        <PostContent content={content} />
+                        {tags && tags.length ? (
+                            <div style={{ marginTop: `4rem` }}>
+                                <h4>Tags</h4>
+                                <ul className="taglist">
+                                    {tags.map((tag) => (
+                                        <li key={tag + `tag`}>
+                                            <Link to={`/tags/${kebabCase(tag)}/`}>{tag}</Link>
+                                        </li>
+                                    ))}
+                                </ul>
+                            </div>
+                        ) : null}
+                    </div>
+                </div>
+            </div>
+        </section>
+    );
+};
+
+interface BlogPostProps {
+    data: { markdownRemark: any };
+}
+
+const BlogPost: React.FC<BlogPostProps> = ({ data }) => {
+    const { markdownRemark: post } = data;
+
+    return (
+        <Layout>
+            <BlogPostTemplate
+                content={post.html}
+                contentComponent={HTMLContent}
+                description={post.frontmatter.description}
+                helmet={
+                    <Helmet titleTemplate="%s | Blog">
+                        <title>{`${post.frontmatter.title}`}</title>
+                        <meta name="description" content={`${post.frontmatter.description}`} />
+                    </Helmet>
+                }
+                tags={post.frontmatter.tags}
+                title={post.frontmatter.title}
+            />
+        </Layout>
+    );
+};
+
+export default BlogPost;
+
+export const pageQuery = graphql`
+    query BlogPostByID($id: String!) {
+        markdownRemark(id: { eq: $id }) {
+            id
+            html
+            frontmatter {
+                date(formatString: "MMMM DD, YYYY")
+                title
+                description
+                tags
+            }
+        }
+    }
+`;
